Refactored TableRecon, Schema, DatabaseConfig Dataclasses to config
