+++ src/databricks/labs/remorph/reconcile/connectors/data_source.py
@@ -4,10 +4,8 @@ from abc import ABC, abstractmethod
 from databricks.sdk import WorkspaceClient  # pylint: disable-next=wrong-import-order
 from pyspark.sql import DataFrame, SparkSession
 
-from databricks.labs.remorph.reconcile.recon_config import (  # pylint: disable=ungrouped-imports
-    JdbcReaderOptions,
-    Schema,
-)
+from databricks.labs.remorph.config import Schema
+from databricks.labs.remorph.reconcile.recon_config import JdbcReaderOptions
 
 
 class DataSource(ABC):
