+++ src/databricks/labs/remorph/reconcile/query_builder/base.py
@@ -1,5 +1,6 @@
 from abc import ABC
 
+from databricks.labs.remorph.config import Schema
 from databricks.labs.remorph.reconcile.connectors.databricks import DatabricksDataSource
 from databricks.labs.remorph.reconcile.connectors.oracle import OracleDataSource
 from databricks.labs.remorph.reconcile.connectors.snowflake import SnowflakeDataSource
@@ -9,7 +10,6 @@ from databricks.labs.remorph.reconcile.constants import (
 )
 from databricks.labs.remorph.reconcile.recon_config import (
     ColumnMapping,
-    Schema,
     Table,
     Transformation,
     TransformRuleMapping,
